Reworked windows changes a bit to localize the changes.
