+++ include/quicly/frame.h
@@ -32,37 +32,7 @@ extern "C" {
 #include "picotls.h"
 #include "quicly/constants.h"
 #include "quicly/ranges.h"
-
-#ifdef __GNUC__
-#define clz(x) __builtin_clz(x)
-#define clzll(x) __builtin_clzll(x)
-#else
-inline int clz(uint32_t x) {
-    unsigned long r = 0;
-    _BitScanReverse(&r, x);
-    FMT_ASSERT(x != 0, "");
-    // Static analysis complains about using uninitialized data
-    // "r", but the only way that can happen is if "x" is 0,
-    // which the callers guarantee to not happen.
-    //FMT_SUPPRESS_MSC_WARNING(6102)
-    return 31 ^ (int)(r);
-}
-
-inline int clzll(uint64_t x) {
-    unsigned long r = 0;
-#  ifdef _WIN64
-    _BitScanReverse64(&r, x);
-#  else
-    // Scan the high 32 bits.
-    if (_BitScanReverse(&r, (uint32_t)(x >> 32))) return 63 ^ (r + 32);
-    // Scan the low 32 bits.
-    _BitScanReverse(&r, (uint32_t)(x));
-#  endif
-    FMT_ASSERT(x != 0, "");
-    //FMT_SUPPRESS_MSC_WARNING(6102)  // Suppress a bogus static analysis warning.
-    return 63 ^ (int)(r);
-}
-#endif
+#include "compat.h"
 
 #define QUICLY_FRAME_TYPE_PADDING 0
 #define QUICLY_FRAME_TYPE_PING 1
