+++ include/quicly/sentmap.h
@@ -255,8 +255,7 @@ int quicly_sentmap__type_packet(quicly_sentmap_t *map, const quicly_sent_packet_
 
 inline void quicly_sentmap_init(quicly_sentmap_t *map)
 {
-    quicly_sentmap_t new_map = {0};
-    *map = new_map;
+    *map = (quicly_sentmap_t){0};
 }
 
 inline int quicly_sentmap_is_open(quicly_sentmap_t *map)
